requirements.txt
streamlit==1.33.0
langchain==0.1.16
langchain-google-genai==1.0.2
pandas==2.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29